requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
dlib==19.24.2
exceptiongroup==1.1.3
filelock==3.13.1
fsspec==2023.12.2
h11==0.14.0
h5py==3.10.0
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.0
imutils==0.5.4
Jinja2==3.1.2
joblib==1.3.2
lazy_loader==0.3
llvmlite==0.41.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
outcome==1.3.0
packaging==23.2
Pillow==10.1.0
PySocks==1.7.1
PyWavelets==1.4.1
regex==2023.10.3
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
selenium==4.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.12
tifffile==2023.7.10
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.0.7
wsproto==1.2.0
zipp==3.17.0
